+++ Tests/SwiftLintFrameworkTests/Resources/FileNameRuleFixtures/Nested.MyType.swift
@@ -1,0 +1,4 @@
+enum Nested {
+    struct MyType {
+    }
+}